Add PrimaryKeyFields to types.Type allowing composite primary keys
Primary keys are not necessarily formed from a single column,
which is assumed by the templates. By adding a new attribute
PrimaryKeyFields that lists all of the fields that make up the
primary key, then it becomes possible to support composite primary
keys by using a custom template. The existing PrimaryKey attribute
is only set when the primary key is composed of a single column,
rather than containing an incorrect value that causes failures.

The existing templates have not yet been modified. If a table
has a composite primary key, code will be generated the same
as if it had no primary key defined rather than broken code
being generated.