Massive overhaul of the query builder and templates
Top to bottom overhaul of the query builder and a clean rewrite of the
query templates. Adds additional --exec option to query mode to enable
"exec only" operations that are not introspected (ie, passed directly to
the templates). Moves all query introspection to the models (via the
--exec flax in gen.sh).

Temporarily breaks the schema templates.