Add support for Dot (GraphViz).
This change brings support to generate a Dot digraph using XO. Currently
the generated graphs are relatively simple, creating a node (styled as
an html table) for each relation (table/view) found in the schema. This
mode does not support query mode at the moment, and may in the future.

Closes #258.