Add full support for stored procedures and functions.
Bar mysql stored procedures with out parameters, this commit brings
support for stored procedures and functions in PostgreSQL, Oracle,
SQLServer and MySQL. SQLite3 does not have user defined procedures. This
change also allows for procs to return multiple types, and distinguishes
between the two with sf_[name] for functions and sp_[name] for
procedures.

This change also fixes a minor issue with sequences for all databases.