go template: use switch-var that is less likely to name collide
The UnmarshalText() method of a Go-Enum generated code used for a
variable switch-statement the name "s".
If the type name started with s like "Salutation", the generated method
receiver name used the same name, also s.

This caused the switch-statement code to break. Both variables that are
accessed in it were called "s", the assignment were done to the
switch-variable instead of to the method receiver var.

Rename the switch var to be called "svar", this should make naming
collisions much more unlikely.