Converting default oracle driver to ora
Latest ora driver supports all the godror features/functionality. As it
is a pure Go driver, changing the preferred/default driver for Oracle to
ora.

Adds a new --go-oracle-type param to the Go template that allows a
setting of either `ora` or `godror` that changes the Go code generated
to the old style. The only change is in how `RETURNING` clauses are used
with database genearted primary IDs (ie sequences) for Oracle.

Removes the _examples/go.{mod,sum} files as they are now redudant since
the examples now use the same database drivers/packages/versions as xo
does.